add sentiment analyzer

+++ db/migrate/20171028222755_create_tweets.rb
@@ -2,7 +2,7 @@ class CreateTweets < ActiveRecord::Migration[5.0]
   def change
     create_table :tweets do |t|
       t.string :content
-      t.decimal :sentiment_score
+      t.float :sentiment_score
       t.datetime :date
 
       t.timestamps
